+++ services/canopy.js
@@ -1,25 +1,23 @@
-const axios = require("axios");
-const { buildCanopyCollection, getValues } = require("./iiif-helpers");
+const {
+  buildCanopyCollection,
+  getBulkManifests,
+  getRootCollection,
+  getValues,
+} = require("./build");
 const fs = require("fs");
 const slugify = require("slugify");
 
-/**
- * directory where canopy instance build data is stored.
- */
-const canopyDirectory = ".canopy";
-
-/**
- *
- * @param {*} env
- */
 module.exports.buildCanopy = (env) => {
-  /**
-   * get root collection and transform it to canopy friendly schema
-   */
   getRootCollection(env.collection).then((json) => {
     /**
-     * make collections
+     * set directory to write canopy structure to
      */
+    const canopyDirectory = ".canopy";
+
+    /**
+     * generate collection data
+     */
+    console.log(`Generating collection data...`);
     const canopyCollection = buildCanopyCollection(json, 0, null);
 
     try {
@@ -43,6 +41,7 @@ module.exports.buildCanopy = (env) => {
     /**
      * create manifest listing
      */
+    console.log(`Creating manifest listing...`);
     const canopyManifests = canopyCollection.items.map((item) => {
       /**
        * what should label look like at this point?
@@ -81,7 +80,9 @@ module.exports.buildCanopy = (env) => {
         }
       }
     );
-    const responses = getBulkManifests(canopyManifests, 10);
+
+    console.log(`Flattening prescribed metadata...`);
+    const responses = getBulkManifests(canopyManifests, 25);
 
     responses.then((manifests) => {
       let canopyMetadata = [];
@@ -119,55 +120,3 @@ module.exports.buildCanopy = (env) => {
     });
   });
 };
-
-const getRootCollection = (id) =>
-  fetch(id)
-    .then((response) => response.text())
-    .then((text) => {
-      try {
-        return JSON.parse(text);
-      } catch (err) {
-        console.warn(err);
-      }
-    })
-    .then((json) => json);
-
-const getBulkManifests = async (items, chunkSize) =>
-  await chunks(
-    items,
-    async (item) => axios.get(item.id).then((result) => result.data),
-    chunkSize
-  );
-
-function all(items, fn) {
-  const promises = items.map((item) => fn(item));
-  return Promise.all(promises);
-}
-
-function series(items, fn) {
-  let result = [];
-  return items
-    .reduce((acc, item) => {
-      acc = acc.then(() => {
-        return fn(item).then((res) => result.push(res));
-      });
-      return acc;
-    }, Promise.resolve())
-    .then(() => result);
-}
-
-function splitToChunks(items, chunkSize = 25) {
-  const result = [];
-  for (let i = 0; i < items.length; i += chunkSize) {
-    result.push(items.slice(i, i + chunkSize));
-  }
-  return result;
-}
-
-function chunks(items, fn, chunkSize = 25) {
-  let result = [];
-  const chunks = splitToChunks(items, chunkSize);
-  return series(chunks, (chunk) => {
-    return all(chunk, fn).then((res) => (result = result.concat(res)));
-  }).then(() => result);
-}
