+++ services/utils.ts
@@ -1,9 +1,2 @@
-/**
- *
- * @param id
- * @returns
- */
 export const getJsonByURI = (uri) =>
-  fetch(uri)
-    .then((response) => response.json())
-    .then((json) => json);
+  fetch(uri).then((response) => response.json());
